refactor(race): Extract race info to widget
- Extracted race info to widget
- Made nullable values in shared type
