+++ packages/client/src/widgets/race/ui/index.ts
@@ -1,1 +1,2 @@
 export { RaceList } from './list';
+export { RaceInfo } from './info';
