+++ packages/client/src/widgets/race/ui/info/index.ts
@@ -1,0 +1,1 @@
+export { RaceInfo } from './RaceInfo';
